% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebro-example-data.R
\docType{data}
\name{cerebro_example}
\alias{cerebro_example}
\title{An example dataset for cerebroViz}
\format{An object of class \code{"matrix"}}
\usage{
cerebro_example
}
\description{
This matrix exmplifies the basic format cerebroViz requires of its input.
}
\keyword{datasets}

