% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroEx.R
\docType{data}
\name{cerebroEx}
\alias{cerebroEx}
\title{An example dataset for cerebroViz}
\format{An object of class \code{"matrix"}}
\usage{
cerebroEx
}
\description{
This matrix exemplifies the basic format cerebroViz requires of its input. The data is raw log2 RPKM expression levels for the POGZ gene taken from the Brainspan developmental transcriptome.
}
\details{
BrainSpan: Atlas of the Developing Human Brain [Internet]. Funded by ARRA Awards 1RC2MH089921-01, 1RC2MH090047-01, and 1RC2MH089929-01. © 2011. Available from: http://developinghumanbrain.org.
}
\keyword{datasets}

