% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroEx2.R
\docType{data}
\name{cerebroEx2}
\alias{cerebroEx2}
\title{An example dataset for cerebroViz}
\format{An object of class \code{"matrix"}}
\usage{
cerebroEx2
}
\description{
This matrix exemplifies the basic format cerebroViz requires of its input. The data is log2 RPKM for the CHD8 gene from the Brainspan developmental transcriptome which has been Z-transformed and LOESS smoothed.
}
\details{
BrainSpan: Atlas of the Developing Human Brain [Internet]. Funded by ARRA Awards 1RC2MH089921-01, 1RC2MH090047-01, and 1RC2MH089929-01. © 2011. Available from: http://developinghumanbrain.org.
}
\keyword{datasets}

