% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{col2hex}
\alias{col2hex}
\title{A function from the 'gplots' package used by cerebroViz() to convert color values to hex values.}
\usage{
col2hex(cname)
}
\arguments{
\item{cname}{color name or value.}
}
\description{
For each input color, convert to a hex value.
}
\examples{
col2hex(cname = "red")
}
\keyword{internal}

