% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{edit.brainCol}
\alias{edit.brainCol}
\title{a function used by cerebroViz() to edit the brain outline and brain background.}
\usage{
\method{edit}{brainCol}(xmlc, brainCol)
}
\description{
for each xml, remove fill attributes from 'brainBackground' & 'brainOutline' and replace them with the designated colors.
}
\examples{
edit.brainCol(xmlc, brainCol)
}
\keyword{brainCol}

