% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{edit.crossHatch}
\alias{edit.crossHatch}
\title{a function used by cerebroViz() to add cross-hatching to regions with missing data.}
\usage{
\method{edit}{crossHatch}(xmlc, tmp)
}
\description{
for each xml, get style and append cross-hatching pattern.
}
\examples{
edit.crossHatch(xmlc)
}
\keyword{internal}

