% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz_06102016_2218.R
\name{edit.legend}
\alias{edit.legend}
\title{a function used by cerebroViz() to edit the legend.}
\usage{
\method{edit}{legend}(xmin, xmed, clamp, xmad, xmax, xmlc, hexVec,
  legend.toggle)
}
\description{
for each xml, map the appropriate colors to the legend
}
\examples{
edit.legend(xmin, xmed, clamp, xmad, xmax, xmlc, hexVec, legend.toggle)
}
\keyword{legend}

