% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{edit.maskReg}
\alias{edit.maskReg}
\title{a function used by cerebroViz() to mask subregions when superior regions are supplied in input data.}
\usage{
\method{edit}{maskReg}(xmlc, x, usrg)
}
\description{
for each superior region, get children region nodes and set opacity to 0.
}
\examples{
edit.maskReg(xmlc,x,usrg)
}
\keyword{usrg}

