% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz_06102016_2218.R
\name{edit.regCol}
\alias{edit.regCol}
\title{a function used by cerebroViz() to map input data to brain regions.}
\usage{
\method{edit}{regCol}(tmp, xmlc, hexVec)
}
\description{
for each xml, add the appropriate fill attributes for each region
}
\examples{
edit.regCol(tmp,xmlc,hexVec)
}
\keyword{regCol}

