% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{edit.svgCol}
\alias{edit.svgCol}
\title{a function used by cerebroViz() to edit the brain outline and brain background.}
\usage{
\method{edit}{svgCol}(xmlc, svgCol)
}
\description{
for each xml, remove fill attributes from 'brainBackground' & 'brainOutline' and replace them with the designated colors.
}
\examples{
edit.svgCol(xmlc, svgCol)
}
\keyword{internal}

