% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{editCrossHatch}
\alias{editCrossHatch}
\title{a function used by cerebroViz() to add cross-hatching to regions with missing data.}
\usage{
editCrossHatch(xmlc, tmp)
}
\description{
for each xml, get style and append cross-hatching pattern.
}
\examples{
editCrossHatch(xmlc)
}
\keyword{internal}

