% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{editLabel}
\alias{editLabel}
\title{a function used by cerebroViz() to add a figure label}
\usage{
editLabel(cXml, timePoint, figLabel, x, indA, regLabel, inpReg)
}
\arguments{
\item{cXml}{list containing the xml object for each SVG.}
}
\description{
for each xml, get label text node a fill with column names or timePoint.
}
\examples{
editLabel(cXml, timePoint, figLabel, x, indA, regLabel, inpReg)
}
\keyword{internal}

