% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{editLegend}
\alias{editLegend}
\title{a function used by cerebroViz() to edit the legend.}
\usage{
editLegend(xmin, xmed, clamp, xmad, xmax, xmlc, regCol, legend.toggle,
  divergent.data)
}
\description{
for each xml, map the appropriate colors to the legend
}
\examples{
editLegend(xmin, xmed, clamp, xmad, xmax, xmlc, regCol, legend.toggle, divergent.data)
}
\keyword{internal}

