% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{editnaHatch}
\alias{editnaHatch}
\title{a function used by cerebroViz() to add cross-hatching to regions with missing data.}
\usage{
editnaHatch(xmlc, lupiter)
}
\arguments{
\item{xmlc}{list containing the xml object for each SVG.}

\item{lupiter}{hex gradient indices for current iteration (timePoint) in the loop, specified within cerebroViz().}
}
\description{
for each xml, get style and append cross-hatching pattern.
}
\examples{
editnaHatch(xmlc, lupiter)
}
\keyword{internal}

