% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{editPal}
\alias{editPal}
\title{a function used by cerebroViz() to map input data to brain regions.}
\usage{
editPal(lupiter, cXml, hexVec, naHatch)
}
\arguments{
\item{lupiter}{hex gradient indices for current iteration (timePoint) in the loop, specified within cerebroViz().}

\item{cXml}{list containing the xml object for each SVG.}

\item{hexVec}{character vector of length 201 containing the hex value gradient for visualization.}

\item{naHatch}{logical indicating if regions of missing data should be filled with a cross-hatch pattern to differentiate them from the brain's background.}
}
\description{
for each xml, add the appropriate fill attributes for each region
}
\examples{
editPal(lupiter, cXml, hexVec, naHatch)
}
\keyword{internal}

