% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{editRegCol}
\alias{editRegCol}
\title{a function used by cerebroViz() to map input data to brain regions.}
\usage{
editRegCol(tmp, xmlc, hexVec, cross.hatch)
}
\description{
for each xml, add the appropriate fill attributes for each region
}
\examples{
editRegCol(tmp,xmlc,hexVec)
}
\keyword{internal}

