% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{editsecondaryPalette}
\alias{editsecondaryPalette}
\title{a function used by cerebroViz() to edit the brain outline and brain background.}
\usage{
editsecondaryPalette(xmlc, secondaryPalette)
}
\arguments{
\item{xmlc}{list containing the xml object for each SVG.}

\item{secondaryPalette}{character vector of length three specifying colors for the brain outline, brain background, and svg background in that order.}
}
\description{
for each xml, remove fill attributes from 'brainBackground' & 'brainOutline' and replace them with the designated colors.
}
\examples{
editsecondaryPalette(xmlc, secondaryPalette)
}
\keyword{internal}

