% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex1-data.R
\docType{data}
\name{ex1}
\alias{ex1}
\title{An example dataset for cerebroViz}
\format{An object of class \code{"matrix"}}
\usage{
ex1
}
\description{
This matrix exmplifies the basic format cerebroViz requires of its input.
}
\keyword{datasets}

