% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex2-data.R
\docType{data}
\name{ex2}
\alias{ex2}
\title{Another example dataset for cerebroViz}
\format{An object of class \code{"matrix"}}
\usage{
ex2
}
\description{
This matrix exmplifies the basic format cerebroViz requires of its input and includes overlapping brian regions.
}
\keyword{datasets}

