% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{maskRegions}
\alias{maskRegions}
\title{A function used by cerebroViz() to lower opacity for superior regions when data is not provided.}
\usage{
maskRegions(cXml, supReg, lupiter)
}
\arguments{
\item{cXml}{list containing the xml object for each SVG.}

\item{supReg}{regions that encompass other brain regions, specified within cerebroViz().}

\item{lupiter}{hex gradient indices for current iteration (timePoint) in the loop, specified within cerebroViz().}
}
\description{
For each missing superior region, set opacity to 0.
}
\examples{
maskRegions(cXml, supReg, lupiter)
}
\keyword{internal}
