% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionMap.R
\docType{data}
\name{regionMap}
\alias{regionMap}
\title{Mapping between regions found in various brain databases & cerebroViz}
\format{An object of class \code{"data.frame"}}
\usage{
regionMap
}
\description{
This dataframe is a reference for the naming conventions of cereBroviz as well as Brainspan, GteX and Epigenomics Roadmap.
}
\keyword{datasets}

