% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{rescale}
\alias{rescale}
\title{A function from the 'gplots' package used by cerebroViz() to convert color values to hex values.}
\usage{
rescale(resc_inp, to = c(0, 1), from = range(resc_inp, na.rm = TRUE, finite
  = TRUE))
}
\arguments{
\item{resc_inp}{numeric vector of values to manipulate.}

\item{to}{output range (numeric vector of length two).}

\item{from}{input range (numeric vector of length two).  If not given, is calculated from the range of ‘resc_inp’.}
}
\description{
For each input color, convert to a hex value.
}
\examples{
rescale(1:100)
}
\keyword{internal}

