% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{unmaskRegions}
\alias{unmaskRegions}
\title{a function used by cerebroViz() to unmask subregions when superior regions are not supplied in input data.}
\usage{
unmaskRegions(xmlc, srg, lupiter)
}
\arguments{
\item{xmlc}{list containing the xml object for each SVG.}

\item{srg}{regions that encompass other brain regions, specified within cerebroViz().}

\item{lupiter}{hex gradient indices for current iteration (timePoint) in the loop, specified within cerebroViz().}
}
\description{
for each missing superior region, set opacity to 0.
}
\examples{
unmaskRegions(xmlc, srg, lupiter)
}
\keyword{internal}

