% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz_06102016_2218.R
\name{cerebroScale}
\alias{cerebroScale}
\title{A data scaling function used by cerebroViz()}
\usage{
cerebroScale(datMat, clamp)
}
\description{
This function scales the data passed to cerebroViz and translates data points to color values.
}
\examples{
cerebroScale(datMat,clamp)
}
\keyword{scale}

