% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{cerebroScale}
\alias{cerebroScale}
\title{A data scaling function used by cerebroViz()}
\usage{
cerebroScale(datMat, clamp, xmed, xmad, divergent.data)
}
\description{
This function scales the data passed to cerebroViz and translates data points to color values.
}
\examples{
cerebroScale(datMat,clamp,xmed,xmad,divergent.data)
}
\keyword{scale}

