% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{cerebroViz}
\alias{cerebroViz}
\title{A tool to visualize biological data mapped to SVG brain diagrams. !!TEST TEXT !! OOPT OOOOTOTOTOTODFV}
\usage{
cerebroViz(x, timepoint = 1, outfile = "cerebroViz_output",
  regCol = c("blue", "grey", "red"), brainCol = c("white", "black"),
  backgroundCol = "white", divergent.data = TRUE, clamp = NULL,
  cross.hatch = FALSE, legend.toggle = TRUE, custom.names = FALSE)
}
\arguments{
\item{x}{a matrix object containing the data to map. Rownames should reflect the appropriate brain region. Columns may represent different time points or replicates.}

\item{timepoint}{a numeric vector of columns in 'x' to visualize.}

\item{outfile}{the desired prefix for the output SVG files.}

\item{regCol}{a character vector of colors to use in the visualization. Accepts color names, hex values, and RGB values. For sequential data, it is recommended to use two colors. Three colors, with a neutral color in the middle, are suggested for divergent data.}

\item{brainCol}{a character vector of length two specifying colors for the outline and background of the brain.}

\item{backgroundCol}{a character vector of length one specifying the color of the SVG background.}

\item{divergent.data}{logical indicating if input data is divergent in nature. Default assumes data is sequential.}

\item{clamp}{a numeric vector of length one specifying the value to multiply by MAD to determine outliers that will be 'clamped' down to prevent skewed visualizations.}

\item{cross.hatch}{logical indicating if regions lacking data should be cross-hatched to differentiate them from the brain's background.}

\item{legend.toggle}{logical indicating if the legend bar should be visible.}

\item{custom.names}{logical indicating if custom naming conventions are used for the input data. If TRUE, user should complete the mapping spreadsheet.}
}
\description{
'cerebroViz' is a data mapping tool for visualizing spatiotemporal data in the brain. The user inputs a matrix of data and the tool outputs publication quality SVG diagrams with color mapping reflective of the input data. 'cerebroViz' supports 30 brain regions used by BrainSpan, GTex, Roadmap Epigneomics, and more.
}
\examples{
x = t(apply(apply(rbind(matrix((sample(c(-400:600),260)/100),nrow=26,ncol=10),matrix(NA,nrow=4,ncol=10)),2,sample),1,sample))
rownames(x) = c("A1C", "ACC", "AMY", "ANG", "BS", "CAU", "CB", "DFC", "FCX", "HIP", "HTH", "IPC", "ITC", "M1C", "MED", "MFC", "OCX", "OFC", "PCX", "PIT", "PUT", "PON", "S1C", "SN", "STC", "STR", "TCX", "THA", "V1C", "VFC")
cerebroViz(x, regCol=c("blue","grey","red"))
}
\keyword{cerebroViz}

