% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{cerebroViz}
\alias{cerebroViz}
\title{A tool to visualize biological data mapped to SVG brain diagrams.}
\usage{
cerebroViz(x, filePrefix = "cerebroViz_output", palette = NULL,
  timePoint = 1, divData = FALSE, secPalette = c("white", "black",
  "white"), clamp = NULL, naHatch = FALSE, legend = TRUE,
  customNames = NULL)
}
\arguments{
\item{x}{matrix containing input data. Row names should reflect the appropriate brain regions. Column names may represent different time points or samples and do not require naming}

\item{filePrefix}{desired prefix for the output SVG files}

\item{palette}{character vector of color values for visualizing brain regions. Accepts color names, hex values, and RGB values}

\item{timePoint}{a numeric vector of columns in 'x' to visualize}

\item{divData}{logical indicating if input data is to be visualized as it diverges from the median. Defaults to false}

\item{secPalette}{character vector of length three specifying colors for the brain background, brain outline/legend, and svg background in that order}

\item{clamp}{coefficient to the Median Absolute Deviation. Added and subtracted from the median to identify a range of non-outliers. Values external to this range will 'clamped' to extremes of the non-outlier range}

\item{naHatch}{logical indicating if regions of missing data should be filled with a cross-hatch pattern to differentiate them from the brain's background}

\item{legend}{logical indicating if the legend bar is displayed}

\item{customNames}{dataframe or matrix with 2 columns. The first column for cerebroViz convention names, the second column for custom user names. Cells in input cannot be factors}
}
\description{
'cerebroViz' is a tool for visualizing spatiotemporal data in the brain within an anatomical context. The user inputs a matrix and the tool creates SVG diagrams with color mapping reflective of the input data. 'cerebroViz' supports 30 brain regions used by BrainSpan, GTEx, Roadmap Epigenomics, and more.
}
\examples{
x = t(apply(apply(rbind(matrix((sample(c(-400:600),260)/100),nrow=26,ncol=10),matrix(NA,nrow=4,ncol=10)),2,sample),1,sample))
rownames(x) = c("A1C", "CNG", "AMY", "ANG", "BS", "CAU", "CB", "DFC", "FL", "HIP", "HTH", "IPC", "ITC", "M1C", "MED", "MFC", "OL", "OFC", "PL", "PIT", "PUT", "PON", "S1C", "SN", "STC", "STR", "TL", "THA", "V1C", "VFC")
cerebroViz(x)
}
\keyword{cerebroViz}

