% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{edit.legend}
\alias{edit.legend}
\title{a function used by cerebroViz() to edit the legend.}
\usage{
\method{edit}{legend}(xmin, xmed, clamp, xmad, xmax, xmlc, regCol,
  legend.toggle, divergent.data)
}
\description{
for each xml, map the appropriate colors to the legend
}
\examples{
edit.legend(xmin, xmed, clamp, xmad, xmax, xmlc, regCol, legend.toggle, divergent.data)
}
\keyword{internal}

