% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{editCrossHatch}
\alias{editCrossHatch}
\title{a function used by cerebroViz() to add cross-hatching to regions with missing data.}
\usage{
editCrossHatch(xmlc, tmp)
}
\arguments{
\item{xmlc}{list containing the xml object for each SVG.}

\item{tmp}{hex gradient indices for current iteration (timepoint) in the loop, specified within cerebroViz().}
}
\description{
for each xml, get style and append cross-hatching pattern.
}
\examples{
editCrossHatch(xmlc, tmp)
}
\keyword{internal}

