% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebroViz.R
\name{unmaskRegions}
\alias{unmaskRegions}
\title{a function used by cerebroViz() to unmask subregions when superior regions are not supplied in input data.}
\usage{
unmaskRegions(xmlc, srg, tmp)
}
\arguments{
\item{xmlc}{xml list for lobe and sagittal, assigned within cerebroViz().}

\item{srg}{regions that encompass other brain regions, specified within cerebroViz().}

\item{tmp}{the color index column for the current timepoint in the loop, specified within cerebroViz().}
}
\description{
for each missing superior region, set opacity to 0.
}
\examples{
unmaskRegions(xmlc,srg,tmp)
}
\keyword{unmaskRegions}

